#613 refactor: Upgrade config-server to ESM
