+++ PowerPi.code-workspace
@@ -208,6 +208,7 @@
             "Common Node Test",
             "Common Python",
             "Common Python Test",
+            "Config Server",
             "Energy Monitor",
             "Event",
             "Scheduler",
