+++ services/config-server/package.json
@@ -3,6 +3,7 @@
     "version": "1.1.8",
     "description": "PowerPi config retrieval from GitHub pushing to message queue",
     "private": true,
+    "type": "module",
     "license": "GPL-3.0-only",
     "author": "TWilkin",
     "repository": {
@@ -23,8 +24,8 @@
     "dependencies": {
         "@octokit/rest": "^20.0.1",
         "@powerpi/common": "^0.2.2",
-        "ajv": "^8.11.0",
-        "ajv-formats": "2.1.1",
+        "ajv": "^8.17.1",
+        "ajv-formats": "3.0.1",
         "ts-command-line-args": "^2.5.1",
         "typedi": "^0.10.0",
         "underscore": "^1.13.4",
