+++ services/config-server/src/schema/common/index.ts
@@ -1,14 +1,14 @@
-import * as AdditionalState from "./AdditionalState.schema.json";
-import * as ColourTemperature from "./ColourTemperature.schema.json";
-import * as Condition from "./Condition.schema.json";
-import * as Cron from "./Cron.schema.json";
-import * as DeviceState from "./DeviceState.schema.json";
-import * as Hue from "./Hue.schema.json";
-import * as Identifier from "./Identifier.schema.json";
-import * as JsonPath from "./JsonPatch.schema.json";
-import * as MACAddress from "./MACAddress.schema.json";
-import * as Percentage from "./Percentage.schema.json";
-import * as Port from "./Port.schema.json";
+import * as AdditionalState from "./AdditionalState.schema.json" with { type: "json" };
+import * as ColourTemperature from "./ColourTemperature.schema.json" with { type: "json" };
+import * as Condition from "./Condition.schema.json" with { type: "json" };
+import * as Cron from "./Cron.schema.json" with { type: "json" };
+import * as DeviceState from "./DeviceState.schema.json" with { type: "json" };
+import * as Hue from "./Hue.schema.json" with { type: "json" };
+import * as Identifier from "./Identifier.schema.json" with { type: "json" };
+import * as JsonPath from "./JsonPatch.schema.json" with { type: "json" };
+import * as MACAddress from "./MACAddress.schema.json" with { type: "json" };
+import * as Percentage from "./Percentage.schema.json" with { type: "json" };
+import * as Port from "./Port.schema.json" with { type: "json" };
 
 export default function loadCommonSchema() {
     return {
