+++ services/config-server/src/schema/config/index.ts
@@ -1,8 +1,8 @@
-import * as devices from "./devices.schema.json";
-import * as events from "./events.schema.json";
-import * as floorplan from "./floorplan.schema.json";
-import * as schedules from "./schedules.schema.json";
-import * as users from "./users.schema.json";
+import * as devices from "./devices.schema.json" with { type: "json" };
+import * as events from "./events.schema.json" with { type: "json" };
+import * as floorplan from "./floorplan.schema.json" with { type: "json" };
+import * as schedules from "./schedules.schema.json" with { type: "json" };
+import * as users from "./users.schema.json" with { type: "json" };
 
 export default function loadConfigSchema() {
     return {
