+++ services/config-server/src/schema/devices/energenie/index.ts
@@ -1,7 +1,7 @@
-import * as EnergenieCommon from "./EnergenieCommon.schema.json";
-import * as EnergeniePairing from "./EnergeniePairing.schema.json";
-import * as EnergenieSocket from "./EnergenieSocket.schema.json";
-import * as EnergenieSocketGroup from "./EnergenieSocketGroup.schema.json";
+import * as EnergenieCommon from "./EnergenieCommon.schema.json" with { type: "json" };
+import * as EnergeniePairing from "./EnergeniePairing.schema.json" with { type: "json" };
+import * as EnergenieSocket from "./EnergenieSocket.schema.json" with { type: "json" };
+import * as EnergenieSocketGroup from "./EnergenieSocketGroup.schema.json" with { type: "json" };
 
 export default function loadEnergenieSchema() {
     return {
