+++ services/config-server/src/schema/devices/harmony/index.ts
@@ -1,5 +1,5 @@
-import * as HarmonyActivity from "./HarmonyActivity.schema.json";
-import * as HarmonyHub from "./HarmonyHub.schema.json";
+import * as HarmonyActivity from "./HarmonyActivity.schema.json" with { type: "json" };
+import * as HarmonyHub from "./HarmonyHub.schema.json" with { type: "json" };
 
 export default function loadHarmonySchema() {
     return {
