+++ services/config-server/src/schema/devices/lifx/index.ts
@@ -1,5 +1,4 @@
-import * as LIFXLight from "./LIFXLight.schema.json";
-
+import * as LIFXLight from "./LIFXLight.schema.json" with { type: "json" };
 export default function loadLIFXSchema() {
     return {
         LIFXLight,
