+++ services/config-server/src/schema/devices/network/index.ts
@@ -1,4 +1,4 @@
-import * as Computer from "./Computer.schema.json";
+import * as Computer from "./Computer.schema.json" with { type: "json" };
 
 export default function loadNetworkSchema() {
     return {
