+++ services/config-server/src/schema/devices/snapcast/index.ts
@@ -1,5 +1,5 @@
-import * as SnapcastClient from "./SnapcastClient.schema.json";
-import * as SnapcastServer from "./SnapcastServer.schema.json";
+import * as SnapcastClient from "./SnapcastClient.schema.json" with { type: "json" };
+import * as SnapcastServer from "./SnapcastServer.schema.json" with { type: "json" };
 
 export default function loadSnapcastSchema() {
     return {
