+++ services/config-server/src/schema/devices/zigbee/index.ts
@@ -1,11 +1,11 @@
-import * as AqaraDoorWindow from "./AqaraDoorWindow.schema.json";
-import * as OsramSwitchMini from "./OsramSwitchMini.schema.json";
-import * as SonoffSwitch from "./SonoffSwitch.schema.json";
-import * as ZigBeeDevice from "./ZigBeeDevice.schema.json";
-import * as ZigBeeEnergyMonitor from "./ZigBeeEnergyMonitor.schema.json";
-import * as ZigBeeLight from "./ZigBeeLight.schema.json";
-import * as ZigBeePairing from "./ZigBeePairing.schema.json";
-import * as ZigBeeSocket from "./ZigBeeSocket.schema.json";
+import * as AqaraDoorWindow from "./AqaraDoorWindow.schema.json" with { type: "json" };
+import * as OsramSwitchMini from "./OsramSwitchMini.schema.json" with { type: "json" };
+import * as SonoffSwitch from "./SonoffSwitch.schema.json" with { type: "json" };
+import * as ZigBeeDevice from "./ZigBeeDevice.schema.json" with { type: "json" };
+import * as ZigBeeEnergyMonitor from "./ZigBeeEnergyMonitor.schema.json" with { type: "json" };
+import * as ZigBeeLight from "./ZigBeeLight.schema.json" with { type: "json" };
+import * as ZigBeePairing from "./ZigBeePairing.schema.json" with { type: "json" };
+import * as ZigBeeSocket from "./ZigBeeSocket.schema.json" with { type: "json" };
 
 export default function loadZigBeeSchema() {
     return {
