+++ services/config-server/src/services/handlers/HandlerFactory.ts
@@ -1,13 +1,13 @@
 import { ConfigFileType, IDeviceConfigFile } from "@powerpi/common";
 import { Service } from "typedi";
-import Container from "../../container";
-import DeviceHandler from "./DeviceHandler";
-import IHandler from "./IHandler";
+import Container from "../../container.js";
+import DeviceHandler from "./DeviceHandler.js";
+import IHandler from "./IHandler.js";
 
 @Service()
 export default class HandlerFactory {
     build<TConfigFile extends IDeviceConfigFile>(
-        type: ConfigFileType
+        type: ConfigFileType,
     ): IHandler<TConfigFile> | undefined {
         switch (type) {
             case ConfigFileType.Devices:
