+++ services/config-server/src/services/tests/devices/EnergenieDevicesConfigFile.test.ts
@@ -1,4 +1,4 @@
-import commonDeviceTests from "./commonDeviceTests";
+import commonDeviceTests from "./commonDeviceTests.js";
 
 describe("Energenie Devices", () => {
     describe("Energenie Pairing", () => {
@@ -10,14 +10,14 @@ describe("Energenie Devices", () => {
             test(`Good timeout ${timeout}`, () =>
                 testValid({
                     devices: [{ type: "energenie_pairing", name: "Pairing", timeout }],
-                }))
+                })),
         );
 
         [-1, 0].forEach((timeout) =>
             test(`Bad timeout ${timeout}`, () =>
                 testInvalid({
                     devices: [{ type: "energenie_pairing", name: "Pairing", timeout }],
-                }))
+                })),
         );
     });
 
@@ -30,14 +30,14 @@ describe("Energenie Devices", () => {
             test(`Good device id ${device_id}`, () =>
                 testValid({
                     devices: [{ type: "energenie_socket", name: "Socket", device_id }],
-                }))
+                })),
         );
 
         [-1, 5].forEach((device_id) =>
             test(`Bad device id ${device_id}`, () =>
                 testInvalid({
                     devices: [{ type: "energenie_socket", name: "Socket", device_id }],
-                }))
+                })),
         );
     });
 
@@ -64,7 +64,7 @@ describe("Energenie Devices", () => {
                             devices: ["Socket"],
                         },
                     ],
-                }))
+                })),
         );
 
         [-1, 16].forEach((home_id) =>
@@ -78,7 +78,7 @@ describe("Energenie Devices", () => {
                             devices: ["Socket"],
                         },
                     ],
-                }))
+                })),
         );
     });
 });
