+++ services/config-server/src/services/tests/devices/SnapcastDevicesConfigFile.test.ts
@@ -1,4 +1,4 @@
-import commonDeviceTests from "./commonDeviceTests";
+import commonDeviceTests from "./commonDeviceTests.js";
 
 describe("Snapcast Devices", () => {
     describe("Client", () => {
