+++ services/config-server/src/services/tests/devices/ZigBeeDevicesConfigFile.test.ts
@@ -1,5 +1,5 @@
-import commonDeviceTests from "./commonDeviceTests";
-import commonSensorTests from "./commonSensorTests";
+import commonDeviceTests from "./commonDeviceTests.js";
+import commonSensorTests from "./commonSensorTests.js";
 
 function commonZigBeeTests(validFile: object, sensor = false) {
     const { getDevice, testValid, testInvalid } = sensor
