+++ services/config-server/src/services/tests/devices/commonSensorTests.ts
@@ -1,10 +1,10 @@
 import { ConfigFileType, IDeviceConfigFile } from "@powerpi/common";
-import ValidatorService from "../../ValidatorService";
+import ValidatorService from "../../ValidatorService.js";
 import {
     testInvalid as _testInvalid,
     testValid as _testValid,
     setupValidator,
-} from "../setupValidator";
+} from "../setupValidator.js";
 
 export default function commonSensorTests(validFile: object) {
     let subject: ValidatorService | undefined;
